feat(client-fluid-framework): expose getPresence from fluid-framework - Move support to new `@fluid-internal/presence-static`, re-export from `fluid-framework`, and deprecate from `@fluidframework/presence` - `fluid-framework` only exports `getPresence*`. Callers can access remaining surface from `@fluidframework/presence`. - Update example and test uses
