+++ examples/apps/presence-tracker/src/app.ts
@@ -3,13 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import {
-	getPresence,
-	type AttendeeId,
-	type AttendeeStatus,
-} from "@fluidframework/presence/beta";
+import type { AttendeeId, AttendeeStatus } from "@fluidframework/presence/beta";
 import { TinyliciousClient } from "@fluidframework/tinylicious-client";
 import type { ContainerSchema, IFluidContainer } from "fluid-framework";
+import { getPresence } from "fluid-framework/beta";
 
 import { FocusTracker } from "./FocusTracker.js";
 import { MouseTracker } from "./MouseTracker.js";
