+++ feeds/internal-test.txt
@@ -39,6 +39,7 @@
 @fluidframework/request-handler
 @fluidframework/react
 @fluidframework/quill-react
+@fluid-internal/presence-static
 @fluid-internal/presence-runtime
 @fluid-internal/presence-definitions
 @fluidframework/presence
