+++ feeds/public.txt
@@ -32,6 +32,7 @@
 @fluidframework/request-handler
 @fluidframework/react
 @fluidframework/quill-react
+@fluid-internal/presence-static
 @fluid-internal/presence-runtime
 @fluid-internal/presence-definitions
 @fluidframework/presence
