+++ packages/framework/fluid-framework/api-report/fluid-framework.legacy.beta.api.md
@@ -306,6 +306,9 @@ export const ForestTypeOptimized: ForestType;
 // @beta
 export const ForestTypeReference: ForestType;
 
+// @beta
+export const getPresence: (fluidContainer: IFluidContainer) => Presence;
+
 // @beta @legacy
 export interface IBranchOrigin {
     id: string;
