+++ packages/framework/presence-static/api-extractor.json
@@ -1,0 +1,8 @@
+{
+	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
+	"extends": "../../../common/build/build-common/api-extractor-base.esm.no-legacy.json",
+	// Disable docModel as customer facing API surface is from `fluid-framework` package.
+	"docModel": {
+		"enabled": false
+	}
+}
