+++ packages/framework/presence-static/src/getPresence.ts
@@ -1,0 +1,42 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import type {
+	Presence,
+	PresenceWithNotifications,
+} from "@fluid-internal/presence-definitions";
+import {
+	ContainerPresenceFactory,
+	extensionId,
+	// eslint-disable-next-line import-x/no-internal-modules -- any import spec is allowed for @fluid-internal/ imports - eslint-config-fluid needs updated
+} from "@fluid-internal/presence-runtime/extension";
+import { assert } from "@fluidframework/core-utils/internal";
+import type { IFluidContainer } from "@fluidframework/fluid-static";
+import { isInternalFluidContainer } from "@fluidframework/fluid-static/internal";
+
+/**
+ * Acquire a {@link @fluidframework/presence#Presence} from a Fluid Container
+ * @param fluidContainer - Fluid Container to acquire the map from
+ * @returns the {@link @fluidframework/presence#Presence}
+ *
+ * @beta
+ */
+export const getPresence: (fluidContainer: IFluidContainer) => Presence = getPresenceAlpha;
+
+/**
+ * Acquire a {@link @fluidframework/presence#PresenceWithNotifications} from a Fluid Container
+ * @param fluidContainer - Fluid Container to acquire the map from
+ * @returns the {@link @fluidframework/presence#PresenceWithNotifications}
+ *
+ * @alpha
+ */
+export function getPresenceAlpha(fluidContainer: IFluidContainer): PresenceWithNotifications {
+	assert(
+		isInternalFluidContainer(fluidContainer),
+		0xa2f /* IFluidContainer was not recognized. Only Containers generated by the Fluid Framework are supported. */,
+	);
+	const presence = fluidContainer.acquireExtension(extensionId, ContainerPresenceFactory);
+	return presence;
+}
