+++ packages/framework/presence-static/tsconfig.json
@@ -1,0 +1,16 @@
+{
+	"extends": "../../../common/build/build-common/tsconfig.node16.json",
+	"include": ["src/**/*"],
+	"exclude": ["src/test/**/*"],
+	"compilerOptions": {
+		"rootDir": "./src",
+		"outDir": "./lib",
+		"noImplicitAny": true,
+		"noImplicitOverride": true,
+		// `skipLibCheck` set to handle improper handling of exactOptionalPropertyTypes
+		// in upstream packages (mainly merge-tree and container-runtime) required via
+		// fluid-static (for isInternalFluidContainer).
+		// Could alternatively build with exactOptionalPropertyTypes: false.
+		"skipLibCheck": true,
+	},
+}
