+++ packages/framework/presence/api-report/presence.alpha.api.md
@@ -49,11 +49,8 @@ export interface BroadcastControlSettings {
 // @beta
 export type ClientConnectionId = string;
 
-// @beta
+// @beta @deprecated
 export const getPresence: (fluidContainer: IFluidContainer) => Presence;
-
-// @alpha
-export function getPresenceAlpha(fluidContainer: IFluidContainer): PresenceWithNotifications;
 
 // @beta @system
 export namespace InternalPresenceTypes {
