+++ packages/framework/presence/src/getPresence.ts
@@ -3,17 +3,14 @@
  * Licensed under the MIT License.
  */
 
-import type {
-	Presence,
-	PresenceWithNotifications,
-} from "@fluid-internal/presence-definitions";
+import type { Presence } from "@fluid-internal/presence-definitions";
 import {
 	ContainerPresenceFactory,
 	extensionId,
 } from "@fluid-internal/presence-runtime/extension";
+import { getPresenceAlpha } from "@fluid-internal/presence-static";
 import { assert } from "@fluidframework/core-utils/internal";
 import type { IFluidContainer } from "@fluidframework/fluid-static";
-import { isInternalFluidContainer } from "@fluidframework/fluid-static/internal";
 import type {
 	FluidDataStoreContextInternal,
 	IFluidDataStoreContext,
@@ -25,24 +22,11 @@ import type {
  * @returns the {@link Presence}
  *
  * @beta
+ *
+ * @deprecated Import from `fluid-framework/beta` instead. This export will be removed in the 2.110.0 release.
+ * See {@link https://github.com/microsoft/FluidFramework/issues/26397}
  */
 export const getPresence: (fluidContainer: IFluidContainer) => Presence = getPresenceAlpha;
-
-/**
- * Acquire a {@link PresenceWithNotifications} from a Fluid Container
- * @param fluidContainer - Fluid Container to acquire the map from
- * @returns the {@link PresenceWithNotifications}
- *
- * @alpha
- */
-export function getPresenceAlpha(fluidContainer: IFluidContainer): PresenceWithNotifications {
-	assert(
-		isInternalFluidContainer(fluidContainer),
-		0xa2f /* IFluidContainer was not recognized. Only Containers generated by the Fluid Framework are supported. */,
-	);
-	const presence = fluidContainer.acquireExtension(extensionId, ContainerPresenceFactory);
-	return presence;
-}
 
 function assertContextHasExtensionProvider(
 	context: IFluidDataStoreContext,
