+++ packages/service-clients/end-to-end-tests/azure-client/src/test/multiprocess/childClient.tool.ts
@@ -17,8 +17,8 @@ import { ConnectionState } from "@fluidframework/container-loader";
 import type { ITelemetryBaseLogger, LogLevel } from "@fluidframework/core-interfaces";
 import type { ScopeType } from "@fluidframework/driver-definitions/legacy";
 import type { ContainerSchema, IFluidContainer } from "@fluidframework/fluid-static";
+import { getPresence } from "@fluidframework/fluid-static/beta";
 import {
-	getPresence,
 	type Attendee,
 	type Presence,
 	StateFactory,
